requirements.txt
# QuantIntel — pip install -r requirements.txt

# LangChain + LangGraph
langchain-core>=0.3.81
langgraph>=0.4.8

# LLM providers (install only what you use)
langchain-anthropic>=0.3.15    # for Anthropic Claude
langchain-openai>=0.3.23       # for OpenAI GPT
langchain-google-genai>=2.1.5  # for Google Gemini

# Market data (free, no API key needed)
yfinance>=0.2.63
stockstats>=0.6.5
pandas>=2.0.0
python-dateutil>=2.8.0

# Memory
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0
